Moved convertBinaryFiles to convert_utils.py
